build($hugo): inline css using readFile template function
BREAKING CHANGE: Forks and users who've customized css need adjustment as shown in the commit message. There's not a whole lot to it, but it's breaking nonetheless.

Using this approach enables the CSS to be moved from the partials directory into the static
directory where it belonged the whole time. Perfect set-up for those who decide to use http2/h2
server push for loading styles more efficiently.

closes #44